Removed out-of-date mfrt.cpp
